[iac] Provision Kueue via KueueAddon with zero-delta adoption
Add KueueAddon, the IaC component that reproduces what
scripts/install_kueue.py --with-queues installs: the cks-kueue Helm
release (webhooks scoped to the Iris namespace), the infiniband and
multinode-nvlink-ib Topology CRs, the cw-ib ResourceFlavor, the
ClusterQueue, the iris-system PriorityClass, and the out-of-band pin of
the kueue-controller-manager to that PriorityClass.

Extract the manifest/value builders and their constants out of
install_kueue.py into a shared importable module,
iris.cluster.platforms.k8s.kueue_manifests, and consume it from both the
script and the IaC component so they render identical manifests. This is
what makes adoption of an already-installed cluster a no-op: previewing
cw-us-east-08a with marin-iac:import=true imports every cluster-scoped
Kueue object with no replace or delete, and the accompanying updates are
benign server-side-apply no-ops (byte-identical specs) plus the
DeploymentPatch re-asserting a field that already holds its value.

The ResourceFlavor's bound topology is now an input (flavor_topology);
GB200 NVL72 binds multinode-nvlink-ib to expose the nvlink.domain level.
Point cw-us-east-08a at the real cluster (marin-us-east-08a) and record
its passphrase-provider encryptionsalt.
