+++ infra/iac/Pulumi.cw-us-east-08a.yaml
@@ -1,2 +1,3 @@
 config:
   marin-iac:cluster: cw-us-east-08a
+encryptionsalt: v1:+eFfBSHNYDA=:v1:slQKtJ+YzGRq4YvD:y2OdKhfS/F0TGyE1pBYCaUZ5Cmo8aQ==
